quant-engine/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
pydantic==2.12.5
pydantic-settings==2.11.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
pandas==2.3.3
numpy==2.4.2
scipy==1.17.1
statsmodels==0.14.5
vectorbt==0.28.4
yfinance==1.2.0
requests==2.32.5
transformers==4.57.1
torch==2.8.0
sentencepiece==0.2.1
python-dateutil==2.9.0.post0
pytz==2026.1.post1
